[Pull Request] `CancellactionToken` support, new `SignalTask` system

+++ Common.BasicHelper.Test/Network/NetUtils_Tests.cs
@@ -8,27 +8,36 @@ public class NetUtils_Tests
     [TestMethod]
     public void Test_IsWebConected()
     {
-        Assert.IsTrue(NetUtils.IsWebConected("localhost", 3));
-        Assert.IsFalse(NetUtils.IsWebConected("192.168.255.255", 3));
+        if (OperatingSystem.IsWindows())
+        {
+            Assert.IsTrue(NetUtils.IsWebConected("localhost", 3));
+            Assert.IsFalse(NetUtils.IsWebConected("192.168.255.255", 3));
+        }
     }
 
     [TestMethod]
     public void Test_DownloadFile()
     {
-        var path = Path.GetFullPath($"{Path.GetTempPath()}/test_downloadFile.txt");
+        if (OperatingSystem.IsWindows())
+        {
+            var path = Path.GetFullPath($"{Path.GetTempPath()}/test_downloadFile.txt");
 
-        NetUtils.DownloadFile(testDownloadFilePath, path);
+            NetUtils.DownloadFile(testDownloadFilePath, path);
 
-        File.Delete(path);
+            File.Delete(path);
+        }
     }
 
     [TestMethod]
     public void Test_WebDownloadFile()
     {
-        var path = Path.GetFullPath($"{Path.GetTempPath()}/test_webDownloadFile.txt");
+        if (OperatingSystem.IsWindows())
+        {
+            var path = Path.GetFullPath($"{Path.GetTempPath()}/test_webDownloadFile.txt");
 
-        NetUtils.WebDownloadFile(testDownloadFilePath, path);
+            NetUtils.WebDownloadFile(testDownloadFilePath, path);
 
-        File.Delete(path);
+            File.Delete(path);
+        }
     }
 }
